Add unused attribute to data parameter
The CTEST2 macros pass in the data parameter to all test cases, setup and
teardown function. However, if the argument is not being used within the
functions the compiler may throw a warning or even error depending on the
activated flags. On gcc for example, the -Wextra flag turns on
-Wunused-parameter which will throw a warning when a parameter is not used.

The UNUSED macro defines __attribute__ ((unused)) and it's being added to data
parameter on setup, teardown and test case macros.

For now it's not available for Windows since I don't think it's compatible with
MVS.

Signed-off-by: Nicolás Pernas Maradei <nicolas.pernas.maradei@emutex.com>